refactor: rename flutter datasource to flutter-version
